pyproject.toml
[project]
name = "utilities"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "msoffcrypto-tool>=5.4.2",
    "pyyaml>=6.0.2",
]
